move lint back
