+++ src/pkg/lint/findings_test.go
@@ -1,8 +1,8 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: 2021-Present The Zarf Authors
 
-// Package rules checks Zarf packages and reports any findings or errors
-package rules
+// Package lint contains functions for verifying zarf yaml files are valid
+package lint
 
 import (
 	"testing"
