+++ src/pkg/lint/lint_test.go
@@ -12,7 +12,6 @@ import (
 	"github.com/stretchr/testify/require"
 	"github.com/zarf-dev/zarf/src/api/v1alpha1"
 	"github.com/zarf-dev/zarf/src/config/lang"
-	"github.com/zarf-dev/zarf/src/pkg/rules"
 	"github.com/zarf-dev/zarf/src/types"
 )
 
@@ -51,13 +50,13 @@ func TestFillComponentTemplate(t *testing.T) {
 
 	findings, err := fillComponentTemplate(&component, createOpts)
 	require.NoError(t, err)
-	expectedFindings := []rules.PackageFinding{
+	expectedFindings := []PackageFinding{
 		{
-			Severity:    rules.SevWarn,
+			Severity:    SevWarn,
 			Description: "There are templates that are not set and won't be evaluated during lint",
 		},
 		{
-			Severity:    rules.SevWarn,
+			Severity:    SevWarn,
 			Description: fmt.Sprintf(lang.PkgValidateTemplateDeprecation, "KEY2", "KEY2", "KEY2"),
 		},
 	}
