+++ src/pkg/packager/generate.go
@@ -16,8 +16,8 @@ import (
 	"github.com/zarf-dev/zarf/src/api/v1alpha1"
 	"github.com/zarf-dev/zarf/src/config"
 	"github.com/zarf-dev/zarf/src/pkg/layout"
+	"github.com/zarf-dev/zarf/src/pkg/lint"
 	"github.com/zarf-dev/zarf/src/pkg/message"
-	"github.com/zarf-dev/zarf/src/pkg/rules"
 )
 
 // Generate generates a Zarf package definition.
@@ -74,7 +74,7 @@ func (p *Packager) Generate(ctx context.Context) (err error) {
 		p.cfg.Pkg.Components[i].Images = images[name]
 	}
 
-	if err := rules.ValidatePackage(p.cfg.Pkg); err != nil {
+	if err := lint.ValidatePackage(p.cfg.Pkg); err != nil {
 		return err
 	}
